requirements.txt
Django==4.2.7
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
prophet==1.1.4
plotly==5.17.0
tensorflow==2.13.0
statsmodels==0.14.0
seaborn==0.12.2
matplotlib==3.7.2
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
redis==5.0.1
celery==5.3.4
django-mathfilters==1.0.0